refactor: organize base sidebar components into dedicated folder
- Create src/components/sidebar/base/ folder structure
- Move BaseSidebar and BaseSidebarHeader to base folder
- Create index.ts for clean exports from base folder
- Update import paths in StorySidebar to use new structure
- Fix formatting and linting issues in base components
- Maintain all existing functionality with improved organization
