+++ src/components/sidebar/base/index.ts
@@ -1,0 +1,2 @@
+export { default as BaseSidebar } from './BaseSidebar';
+export { default as BaseSidebarHeader } from './BaseSidebarHeader';
